Update structs and tests to store all 256 bits of sha256 hash

+++ src/message.h
@@ -34,14 +34,14 @@ typedef struct peer_message {
 typedef struct ask_message {
   uint32_t message_size; // sizeof(ask_message) - 4
   uint8_t type;          // should be set to 0
-  uint32_t sha256[4];
+  uint64_t sha256[4];
 } ask_message;
 
 // a struct for describing the info needed to send a 1MB piece
 typedef struct give_message {
   uint32_t message_size; // sizeof(give_message) - 4
   uint8_t type;          // should be set to 1
-  uint32_t sha256[4];
+  uint64_t sha256[4];
   uint8_t piece[PIECE_SIZE_BYTES]; // 1 MiB
 } give_message;
